fix: rename files for consistency
