+++ native-client/src/navigation/SignedInNavigator.js
